requirements.txt
Django==1.7.1
Pillow==2.7.0
django-mptt==0.6.1
urllib3==1.10
psycopg2==2.5.3
